+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Next version
+
+## Features
+
+- Updated `@dfinity/nns` to add support to a new type of action `RegisterExtension`.
+- Updated `@dfinity/nns` to add support for setting following for multiple topics (`setFollowing`).
+
 # v70
 
 ## Overview